+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - \_process_json:  Added gen_libs.print_data call for JSON format.
 - \_process_json:  Added suppress standard out option.
 - call_run_chk:  Added suppress standard out option to function calls.
+- Documentation update.
 
 ### Removed
 - setup_mail:  Replaced by gen_class.setup_mail call.